+++ src/main/java/com/beginvegan/domain/review/dto/response/ReviewListRes.java
@@ -1,4 +1,4 @@
-package com.beginvegan.domain.review.dto;
+package com.beginvegan.domain.review.dto.response;
 
 import lombok.Builder;
 import lombok.Data;
